requirements.txt
azure-cognitiveservices-vision-computervision
msrest
pillow
streamlit